olpc/sdhci: add /pwrseq0 power sequencer for wlan
This creates a MMC power sequencer instance that fiddles the reset and
power pins in order to ensure the card is initialized and powered up
when necessary. It's defined by the mmc/mmc-pwrseq-sd8787.txt binding.
Link it to MMC controller WLAN card is connected to following the
mmc/mmc.txt bindings.