mmp3: tidy up interrupt controller references
Drop unnecessary interrupt-parent.

Qualify the names with the bus location so that it's obvious which
interrupt controller do we mean (there also are muxes and possibly a GIC).

Don't include INTC in MUX paths. It is there by accident -- the MUXes
cna happily multiplex interrupts via the GIC as well. This will make it
easier to tidy things up.