mmp3/gic: create the TWD nodes
Arm TWD is used for local timer interrupt delivery on SMP systems. Enable it
and create the appropriate nodes in the GIC fixup path, because it relies on
GIC.

Apparently the non-GIC interrupt controller can be used for CPU local
timer interrupt delivery, but not with mainline Linux at this point.