video/dacs: add driver for Chrontel CH7033
This is a driver for video encoder with VGA and DVI/HDMI outputs.

There is no documentation for the chip -- the operation was guessed from
what was sniffed on a Dell Wyse 3020 ThinOS terminal with some clues from
largely obfuscated Mediatek's GPL code dump.

The configuration is hardcoded to what makes sense for 1024x768@60.
However, with VGA and clock bypasses on, it doesn't seem to matter too
much and other modes seem to work fine, which is good.