olpc/4.0: optionally build DT for OLPC's 3.5 kernel
Many of the preceding changes were done for the sake of compatibility
with mainline kernel and were done without considering compatibility
with OLPC XO-4.0's kernel, based on linux-3.5.

Most notably, the clock driver uses completely different DT bindings.
There are a couple more things -- some drivers are crashy when they
don't find the expected nodes, such as I2C node aliases or dumb panel.

This adds an "olpc-compat-" drop-in whose purpose is to beat the tree
back into shape that's acceptable for the non-mainline kernel.
It's loaded just prior to booting the kernel if olpc-compat? flag is
set.